+++ buession-web/src/main/java/com/buession/web/reactive/aop/handler/ReactiveDocumentMetaDataAnnotationHandler.java
@@ -24,34 +24,15 @@
  */
 package com.buession.web.reactive.aop.handler;
 
-import com.buession.aop.MethodInvocation;
 import com.buession.web.aop.handler.AbstractDocumentMetaDataAnnotationHandler;
-import com.buession.web.mvc.view.document.DocumentMetaData;
-import com.buession.web.reactive.aop.AopUtils;
-import com.buession.web.reactive.http.ServerHttp;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * @author Yong.Teng
  */
 public class ReactiveDocumentMetaDataAnnotationHandler extends AbstractDocumentMetaDataAnnotationHandler {
 
-	private final static Logger logger = LoggerFactory.getLogger(ReactiveDocumentMetaDataAnnotationHandler.class);
-
 	public ReactiveDocumentMetaDataAnnotationHandler(){
 		super();
-	}
-
-	@Override
-	public void execute(MethodInvocation mi, DocumentMetaData documentMetaData){
-		ServerHttp serverHttp = AopUtils.getServerHttp(mi);
-		if(serverHttp == null || serverHttp.getModel() == null){
-			logger.debug("{} is null.", serverHttp == null ? "ServerHttp" : "Model");
-			return;
-		}
-
-		addModelAttribute(serverHttp.getModel(), documentMetaData);
 	}
 
 }